tabular_review/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.4
deprecation==2.1.0
dill==0.4.0
dnspython==2.7.0
docling==2.36.1
docling-core==2.37.0
docling-ibm-models==3.4.4
docling-parse==4.0.5
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.2.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.0
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
jinja2==3.1.6
jose==1.0.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
latex2mathml==3.78.0
lazy-loader==0.4
lxml==5.4.0
mako==1.3.10
markdown-it-py==3.0.0
marko==2.1.4
markupsafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.18
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pypdfium2==4.30.1
pytest==8.4.1
pytest-mock==3.14.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
realtime==2.4.3
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
semchunk==2.2.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.5.10
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
xlsxwriter==3.2.3
yarl==1.20.1
